requirements.txt
absl-py==2.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.6.15
cffi==1.17.1
cryptography==45.0.4
deprecation==2.1.0
exceptiongroup==1.3.0
flatbuffers==25.2.10
frozenlist==1.7.0
gast==0.6.0
google-pasta==0.2.0
gotrue==2.12.0
gunicorn==20.1.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html5tagger==1.3.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jwt==1.3.1
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.4
namex==0.1.0
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pip-autoremove==0.10.0
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.2
protobuf==5.29.5
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
realtime==2.4.3
rich==14.0.0
sanic==25.3.0
sanic-routing==23.12.0
six==1.17.0
sniffio==1.3.1
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.3
supafunc==0.9.4
termcolor==3.1.0
tomli==2.2.1
tracerite==1.1.1
typing-inspection==0.4.1
typing_extensions==4.14.0
ujson==5.10.0
uvloop==0.21.0
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.1
